requirements/base.txt
Django>=4.2,<4.3
python-dotenv==1.0.1
django-tailwind==3.8.0
pillow==10.2.0
whitenoise==6.6.0
django-solo==2.2.0
django-inline-svg==0.1.1
djangorestframework==3.15.1
django-cors-headers==4.3.1
coreapi==2.3.3
